Moved the Back button in the settings screen to an objectively superior location.
SavedPreferences now works. The Reset button in the settings is wired to reset the highscore to 100. May need to be changed in the future.

Grid size now scales depending on the selected size. Shout outs to my boy Jason for pointing out where I was messing up in my code (I was making a rookie object instantiation mistake).

The ball is in your court Kenney, let me know if you need help with anything.